[refactor] refactor code
