+++ erpnext_shopify/erpnext_shopify/doctype/shopify_error_log/shopify_error_log.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2015, Frappe Technologies Pvt. Ltd. and contributors
+# For license information, please see license.txt
+
+from __future__ import unicode_literals
+import frappe
+from frappe.model.document import Document
+
+class ShopifyErrorLog(Document):
+	pass
