+++ erpnext_shopify/erpnext_shopify/doctype/shopify_error_log/test_shopify_error_log.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2015, Frappe Technologies Pvt. Ltd. and Contributors
+# See license.txt
+from __future__ import unicode_literals
+
+import frappe
+import unittest
+
+# test_records = frappe.get_test_records('Shopify Error Log')
+
+class TestShopifyErrorLog(unittest.TestCase):
+	pass
