+++ erpnext_shopify/erpnext_shopify/doctype/shopify_settings/shopify_settings.js
@@ -37,7 +37,7 @@ frappe.ui.form.on("Shopify Settings", "refresh", function(frm){
 		cur_frm.add_custom_button(__('Sync Shopify'),
 			function() {  
 				frappe.call({
-					method:"erpnext_shopify.erpnext_shopify.doctype.shopify_settings.shopify_settings.sync_shopify",
+					method:"erpnext_shopify.api.sync_shopify",
 					freeze: true,
 					freeze_message: "Syncing",
 					callback:function(r){
